+++ src/utils/helper.js
@@ -36,10 +36,11 @@ async function removeEmptyAccounts(indexAddress, keyringInstance, vaultState, un
   let accountsArray = [];
   accountsArray.push({ address: indexAddress, isDeleted: false, isImported: false, label: 'Wallet 1' });
   let labelCounter = 2;  // as an initial wallet is already created above with label 'Wallet 1'
+  const chains = Object.keys(Chains.evmChains);
 
   if( recoverMechanism === 'logs'){
     for(let i=0; i < logs.length; i++){
-      if (logs[i].action === 'add-account' && logs[i].chain === "ethereum"){
+      if (logs[i].action === 'add-account' && (chains.includes(logs[i].chain) || logs[i].chain === undefined)){
         const vaultState = await keyringInstance.addNewAccount(keyring[0]);
         const newAccountAddr = Web3.utils.toChecksumAddress(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1])
         if (Web3.utils.toChecksumAddress(logs[i].address) === newAccountAddr) {